MDEV-31122 Server crash in get_lock_data / mysql_lock_abort_for_thread
While ALTER thread tries to notify SELECT thread about lock conflict
it accesses its TABLE object (THD::notify_shared_lock()) and lock data
(mysql_lock_abort_for_thread()). As part of accessing lock data it
calls ha_partition::store_lock() which iterates over all partitions
and does their store_lock().

The problem is SELECT opened 2 read partitions, but
ha_partition::store_lock() tries to access all partitions as indicated
in m_tot_parts which is 4. So the last 2 partitions m_file[2] and
m_file[3] are uninitialized and store_lock() accesses uninitialized
data.

The code in ha_partition::store_lock() does this wrong handling to use
all partitions specifically for the case of
mysql_lock_abort_for_thread(), this is conducted with comment:

  /*
    This can be called from get_lock_data() in mysql_lock_abort_for_thread(),
    even when thd != table->in_use. In that case don't use partition pruning,
    but use all partitions instead to avoid using another threads structures.
  */
  if (thd != table->in_use)
  {
    for (i= 0; i < m_tot_parts; i++)
      to= m_file[i]->store_lock(thd, to, lock_type);
  }

The explanation is "to avoid using another threads structures" does
not really explain why this change was needed.

The change was originally introduced by:

commit 9b7cccaf319
Author: Mattias Jonsson <mattias.jonsson@oracle.com>
Date:   Wed May 30 00:14:39 2012 +0200

    WL#4443:
    final code change for dlenevs review.
    - Don't use pruning in lock_count().
    - Don't use pruning in store_lock() if not owning thd.
    - Renamed is_fields_used_in_trigger to
      is_fields_updated_in_trigger() and check if they
      may be updated.
    - moved out mark_fields_used(TRG_EVENT_UPDATE)
      from mark_columns_needed_for_update().
      And reverted the changed call order. And call
      mark_fields_used(TRG_EVENT_UPDATE) instead.

which also fails to explain the rationale of the change. The original
idea of WL#4443 is to reduce locks and this change does not happen to
serve this goal.

So reverting this change restores original behaviour of using only
partitions marked for use and fixes invalid access to uninitialized
data.